WS does not crash program
Errors send a log, and if it connects and disconnects the program is still good. Test starting the terminal, then the backend, wait for the heartbeat to go through, then close the backend to make sure the program still is able to survive.

Potentially if you close the backend at the exact millisecond that the if statement verifies the WS is still open, then you could still crash the program when it tries to send the message.